auth_service/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.4.26
#click==8.2.1
fastapi==0.115.12
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
mypy==1.16.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
pydantic==2.11.5
pydantic[email]
pydantic_core==2.33.2
pytest==8.3.5
pytest-asyncio==1.0.0
PyYAML==6.0.2
ruff==0.11.12
sniffio==1.3.1
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.3
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
asyncpg==0.30.0
greenlet==3.2.3
python-dotenv== 1.1.1
alembic==1.16.2
python-jose==3.5.0
passlib==1.7.4
python-multipart==0.0.20
bcrypt==4.3.0